biomaRt: iterate through mirrors upon error
With the current configuration, it will iterate through the whole set of mirrors three times and then finally give up. At the end of each round, the script will wait for 30 seconds in the hope that intermittent errors go away. And the script explicitly does not start with the default mirror "www", as this is the one with very frequent errors.