fix: simpler three prime QuantSeq cutadapt setup (#78)
I have tested this on actual QuantSeq data, and the cutadapt setup in
this PR yields mostly the same results that the complicated previous
three step setup recommended by Lexogen gave. It seems to be a bit more
aggressive in the poly-A tail removal, but this should not affect
results much. And it greatly simplifies the workflow setup.