fix: move summing of go-term fold changes out of log-space (#141)
As [discussed in a separate pull
request](https://github.com/snakemake-workflows/rna-seq-kallisto-sleuth/pull/136#discussion_r1983484603),
it seems like we are currently multiplying fold changes across the genes
within a GO term, as we are adding up `log2()` transformations of the
fold changes. To get a sum instead of a product, we have to first take
`2^<log2()-value>`. And if we want to end up with a log2-value of the
sum, we afterwards have to take the `log2()` again. This is what this
pull request suggests.

Not sure though, if this is exactly what we want. Two things we might do
differently are:
1. We might want to leave the fold change in non-log space. But thinking
about this again, we need log2 fold changes for the (signed) pi value
calculation, right?
2. We might want to normalize for the number of genes in a GO term.
Otherwise, a ranking of GO terms based on the signed pi value will
easily be dominated by high level GO terms that simply encompass a lot
of genes. I'll put in a review comment with the respective suggestion.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

## Summary by CodeRabbit

- **Refactor**
- Updated the calculation logic in the analysis process to deliver more
nuanced and refined aggregated metrics.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>
Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>